+++ examples/cpp_regions/supported_pointer.cpp
@@ -1,0 +1,7 @@
+#include <cstddef>
+
+void transform(float *dst, const float *src, std::size_t n) noexcept {
+    for (std::size_t i = 0; i < n; ++i) {
+        dst[i] = src[i] * src[i] + 0.25f;
+    }
+}